1. add TextureReader config macro define 2. fix some size_t code

+++ src/apple/platform_helper_apple.mm
@@ -57,7 +57,7 @@ namespace ERI {
 		NSMutableArray* keys = [[frames allKeys] mutableCopy];
 		[keys sortUsingSelector:@selector(compare:)];
 
-		int num = [keys count];
+		NSInteger num = [keys count];
 		for (int i = 0; i < num; ++i)
 		{
 			unit = [frames objectForKey:[keys objectAtIndex:i]];
